Move all -css.js files to new client-src/css dir.
