+++ client-src/elements/chromedash-feature-filter.js
@@ -1,5 +1,5 @@
 import {LitElement, css, html, nothing} from 'lit';
-import {SHARED_STYLES} from '../sass/shared-css.js';
+import {SHARED_STYLES} from '../css/shared-css.js';
 import {
   FEATURE_CATEGORIES,
   FEATURE_TYPES,
