+++ client-src/elements/chromedash-feature-row.js
@@ -1,5 +1,5 @@
 import {LitElement, css, html, nothing} from 'lit';
-import {SHARED_STYLES} from '../sass/shared-css.js';
+import {SHARED_STYLES} from '../css/shared-css.js';
 import {
   GATE_TEAM_ORDER,
   STAGE_SHORT_NAMES,
