+++ client-src/elements/chromedash-guide-verify-accuracy-page.js
@@ -9,8 +9,8 @@ import {formatFeatureForEdit,
   VERIFY_ACCURACY_METADATA_FIELDS,
   VERIFY_ACCURACY_TRIAL_EXTENSION_FIELDS} from './form-definition';
 import {STAGE_SHORT_NAMES, STAGE_SPECIFIC_FIELDS} from './form-field-enums.js';
-import {SHARED_STYLES} from '../sass/shared-css.js';
-import {FORM_STYLES} from '../sass/forms-css.js';
+import {SHARED_STYLES} from '../css/shared-css.js';
+import {FORM_STYLES} from '../css/forms-css.js';
 
 
 export class ChromedashGuideVerifyAccuracyPage extends LitElement {
