+++ client-src/elements/chromedash-settings-page.js
@@ -1,7 +1,7 @@
 import {LitElement, css, html, nothing} from 'lit';
 import {showToastMessage} from './utils';
-import {SHARED_STYLES} from '../sass/shared-css.js';
-import {FORM_STYLES} from '../sass/forms-css.js';
+import {SHARED_STYLES} from '../css/shared-css.js';
+import {FORM_STYLES} from '../css/forms-css.js';
 
 
 export class ChromedashSettingsPage extends LitElement {
