+++ client-src/elements/chromedash-stack-rank-page.js
@@ -1,7 +1,7 @@
 import {LitElement, css, html} from 'lit';
 import './chromedash-stack-rank';
 import {showToastMessage} from './utils';
-import {SHARED_STYLES} from '../sass/shared-css.js';
+import {SHARED_STYLES} from '../css/shared-css.js';
 
 
 export class ChromedashStackRankPage extends LitElement {
