+++ client-src/elements/chromedash-toast.js
@@ -1,5 +1,5 @@
 import {LitElement, css, html} from 'lit';
-import {SHARED_STYLES} from '../sass/shared-css.js';
+import {SHARED_STYLES} from '../css/shared-css.js';
 
 const DEFAULT_DURATION = 7000;
