+++ developer-documentation.md
@@ -38,7 +38,7 @@ All the pages are rendered in a combination of Jinja2 template (`/templates`) an
    - We pass backend variables to js like this: `const variableInJs = {{variable_in_template|safe}}`.
 1. All Lit components are in `/client-src/elements`.
 1. All JavaScript files are in `/client-src/js-src/` and processed by gulp, then output to `/static/js/` and get included in templates.
-2. All `*-css.js` files used in client-side components are in `/client-src/sass/`.  The remaining css files still being included in templates are in `/static/css/`.
+2. All `*-css.js` files used in client-side components are in `/client-src/css/`.  The remaining css files still being included in templates are in `/static/css/`.
 
 ### Adding an icon
